move setup_connection
