+++ lmfdb/backend/database.py
@@ -16,12 +16,30 @@ from glob import glob
 from psycopg2 import connect, DatabaseError
 from psycopg2.sql import SQL, Identifier, Placeholder
 
-from .encoding import setup_connection, Json
+from .encoding import Json
 from .base import PostgresBase, _meta_tables_cols
 from .table import PostgresTable
 from .utils import DelayCommit
 
 
+def setup_connection(conn):
+    # We want to use unicode everywhere
+    register_type(UNICODE, conn)
+    register_type(UNICODEARRAY, conn)
+    conn.set_client_encoding('UTF8')
+    cur = conn.cursor()
+    cur.execute("SELECT NULL::numeric")
+    oid = cur.description[0][1]
+    NUMERIC = new_type((oid,), "NUMERIC", numeric_converter)
+    cur.execute("SELECT NULL::numeric[]")
+    oid = cur.description[0][1]
+    NUMERICL = new_array_type((oid,), "NUMERIC[]", NUMERIC)
+    register_type(NUMERIC, conn)
+    register_type(NUMERICL, conn)
+    register_adapter(Integer, AsIs)
+    register_adapter(RealNumber, RealEncoder)
+    register_adapter(dict, Json)
+    register_json(conn, loads=Json.loads)
 
 class PostgresDatabase(PostgresBase):
     """
